Add the option to keep temp directory alive at python exit

+++ src/ansys/pytwin/settings.py
@@ -139,7 +139,7 @@ def modify_pytwin_logging(
     PYTWIN_SETTINGS.modify_logging(new_option=new_option, new_level=new_level)
 
 
-def modify_pytwin_working_dir(new_path: str, erase: bool = True):
+def modify_pytwin_working_dir(new_path: str, erase: bool = True, cleanup_temp_dir: bool = True):
     """
     Modify the global PyTwin working directory.
 
@@ -147,11 +147,17 @@ def modify_pytwin_working_dir(new_path: str, erase: bool = True):
     ----------
     new_path: str
         Absolute path to the working directory to use for PyTwin. The directory is created if it does not exist.
+        This directory won't be cleanup at python process exit.
     erase: bool, optional
         Whether to erase a non-empty existing working directory. The default is ``True``,
         in which case the existing working directory is erased and a new one is created.
         If ``False``, the existing working directory is used as it is. This parameter has no
         effect if the directory does not exist.
+    cleanup_temp_dir: bool, optional
+        By default, PyTwin temporary directory is automatically cleaned up at exit of the python process
+        that imported pytwin.
+        When this option is set to False, then PyTwin temporary directory won't be deleted at python process exit.
+        Note that modifying the default behavior may lead to an overflow of the temporary directory.
 
     Raises
     ------
@@ -197,7 +203,7 @@ def modify_pytwin_working_dir(new_path: str, erase: bool = True):
 
     _check_wd_path_is_valid(new_path)
     _check_wd_erase_is_valid(erase)
-    PYTWIN_SETTINGS.modify_wd_dir(new_path=new_path, erase=erase)
+    PYTWIN_SETTINGS.modify_wd_dir(new_path=new_path, erase=erase, cleanup_temp_dir=cleanup_temp_dir)
 
 
 def pytwin_logging_is_enabled():
@@ -265,6 +271,7 @@ class _PyTwinSettings(object):
     SESSION_ID = None
     WORKING_DIRECTORY_PATH = None
     TEMP_WORKING_DIRECTORY_PATH = None
+    CLEANUP_TEMP_WD_AT_EXIT = True
 
     # Immutable constants
     LOGGER_NAME = "pytwin_logger"
@@ -415,7 +422,7 @@ class _PyTwinSettings(object):
                         shutil.copyfile(os.path.join(root, file), os.path.join(dest_path, file))
 
     @staticmethod
-    def modify_wd_dir(new_path: str, erase: bool):
+    def modify_wd_dir(new_path: str, erase: bool, cleanup_temp_dir: bool):
         old_path = _PyTwinSettings.WORKING_DIRECTORY_PATH
 
         # Check new directory
@@ -432,6 +439,7 @@ class _PyTwinSettings(object):
         _PyTwinSettings.WORKING_DIRECTORY_PATH = new_path
         if old_path is not None:
             _PyTwinSettings._migration_due_to_new_wd(old_path, new_path)
+        _PyTwinSettings.CLEANUP_TEMP_WD_AT_EXIT = cleanup_temp_dir
 
     @staticmethod
     def modify_logging(new_option: PyTwinLogOption, new_level: PyTwinLogLevel):
@@ -473,7 +481,8 @@ def cleanup_temp_pytwin_working_directory():
     pytwin_logger.info(PYTWIN_SETTINGS.PYTWIN_END_MSG)
     pytwin_logger.handlers.clear()
     try:
-        shutil.rmtree(PYTWIN_SETTINGS.TEMP_WORKING_DIRECTORY_PATH)
+        if PYTWIN_SETTINGS.CLEANUP_TEMP_WD_AT_EXIT:
+            shutil.rmtree(PYTWIN_SETTINGS.TEMP_WORKING_DIRECTORY_PATH)
     except BaseException as e:
         msg = "Something went wrong while trying to cleanup pytwin temporary directory!"
         msg += f"error message:\n{str(e)}"
